add privacy folder
