+++ src/synthcity/plugins/core/constraints.py
@@ -159,7 +159,7 @@ class Constraints(BaseModel):
                 thresh,
             )
             if res.sum() < prev:
-                log.error(
+                log.info(
                     f"[{feature}] quality loss for constraints {op} = {thresh}. Remaining {res.sum()}. prev length {prev}. Original dtype {X[feature].dtype}.",
                 )
         return res
