+++ src/synthcity/plugins/privacy/plugin_privbayes.py
@@ -598,7 +598,7 @@ class PrivBayesPlugin(Plugin):
 
     @staticmethod
     def type() -> str:
-        return "generic"
+        return "privacy"
 
     @staticmethod
     def hyperparameter_space(**kwargs: Any) -> List[Distribution]:
