+++ tests/plugins/privacy/test_pategan.py
@@ -1,11 +1,8 @@
-# stdlib
-import sys
-
 # third party
 import numpy as np
 import pandas as pd
 import pytest
-from generic_helpers import generate_fixtures, get_airfoil_dataset
+from fhelpers import generate_fixtures, get_airfoil_dataset
 from sklearn.datasets import load_iris
 
 # synthcity absolute
@@ -13,9 +10,14 @@ from synthcity.metrics.eval import PerformanceEvaluatorXGB
 from synthcity.plugins import Plugin
 from synthcity.plugins.core.constraints import Constraints
 from synthcity.plugins.core.dataloader import GenericDataLoader
-from synthcity.plugins.generic.plugin_adsgan import plugin
+from synthcity.plugins.privacy.plugin_pategan import plugin
 
-plugin_name = "adsgan"
+plugin_name = "pategan"
+plugin_args = {
+    "generator_n_layers_hidden": 1,
+    "generator_n_units_hidden": 10,
+    "lamda": 0.1,
+}
 
 
 @pytest.mark.parametrize("test_plugin", generate_fixtures(plugin_name, plugin))
@@ -30,73 +32,42 @@ def test_plugin_name(test_plugin: Plugin) -> None:
 
 @pytest.mark.parametrize("test_plugin", generate_fixtures(plugin_name, plugin))
 def test_plugin_type(test_plugin: Plugin) -> None:
-    assert test_plugin.type() == "generic"
+    assert test_plugin.type() == "privacy"
 
 
 @pytest.mark.parametrize("test_plugin", generate_fixtures(plugin_name, plugin))
 def test_plugin_hyperparams(test_plugin: Plugin) -> None:
-    assert len(test_plugin.hyperparameter_space()) == 11
+    assert len(test_plugin.hyperparameter_space()) == 20
 
 
-def test_plugin_fit() -> None:
-    test_plugin = plugin(
-        n_iter=100, generator_n_layers_hidden=1, generator_n_units_hidden=10
-    )
-
-    df = pd.DataFrame(load_iris()["data"])
-    X = GenericDataLoader(df)
-
-    test_plugin.fit(X)
+@pytest.mark.parametrize(
+    "test_plugin", generate_fixtures(plugin_name, plugin, plugin_args)
+)
+def test_plugin_fit(test_plugin: Plugin) -> None:
+    X = pd.DataFrame(load_iris()["data"])
+    test_plugin.fit(GenericDataLoader(X))
 
 
-def test_plugin_generate() -> None:
-    test_plugin = plugin(
-        n_iter=100, generator_n_layers_hidden=1, generator_n_units_hidden=10
-    )
+def test_plugin_generate_pategan() -> None:
+    test_plugin = plugin(**plugin_args)
+    X = get_airfoil_dataset()
 
-    df = get_airfoil_dataset()
-    X = GenericDataLoader(df)
-
-    test_plugin.fit(X)
+    test_plugin.fit(GenericDataLoader(X))
 
     X_gen = test_plugin.generate()
     assert len(X_gen) == len(X)
-    assert X_gen.shape[1] == df.shape[1]
     assert test_plugin.schema_includes(X_gen)
+    assert X_gen.shape[1] == X.shape[1]
 
     X_gen = test_plugin.generate(50)
     assert len(X_gen) == 50
     assert test_plugin.schema_includes(X_gen)
 
 
-@pytest.mark.skipif(sys.platform != "linux", reason="Linux only for faster results")
-def test_plugin_conditional_adsgan() -> None:
-    test_plugin = plugin(generator_n_units_hidden=5)
-    Xraw, y = load_iris(as_frame=True, return_X_y=True)
-    Xraw["target"] = y
-
-    X = GenericDataLoader(Xraw)
-    test_plugin.fit(X, cond=y)
-
-    X_gen = test_plugin.generate(2 * len(X))
-    assert len(X_gen) == 2 * len(X)
-    assert test_plugin.schema_includes(X_gen)
-
-    count = 10
-    X_gen = test_plugin.generate(count, cond=np.ones(count))
-    assert len(X_gen) == count
-
-    assert (X_gen["target"] == 1).sum() >= 0.8 * count
-
-
 def test_plugin_generate_constraints() -> None:
-    test_plugin = plugin(
-        n_iter=100, generator_n_layers_hidden=1, generator_n_units_hidden=10
-    )
-
-    Xraw = pd.DataFrame(load_iris()["data"])
-    X = GenericDataLoader(Xraw)
-    test_plugin.fit(X)
+    test_plugin = plugin(**plugin_args)
+    X = pd.DataFrame(load_iris()["data"])
+    test_plugin.fit(GenericDataLoader(X))
 
     constraints = Constraints(
         rules=[
@@ -111,15 +82,15 @@ def test_plugin_generate_constraints() -> None:
         ]
     )
 
-    X_gen = test_plugin.generate(constraints=constraints)
+    X_gen = test_plugin.generate(constraints=constraints).dataframe()
     assert len(X_gen) == len(X)
     assert test_plugin.schema_includes(X_gen)
-    assert constraints.filter(X_gen.dataframe()).sum() == len(X_gen)
+    assert constraints.filter(X_gen).sum() == len(X_gen)
 
-    X_gen = test_plugin.generate(count=50, constraints=constraints)
+    X_gen = test_plugin.generate(count=50, constraints=constraints).dataframe()
     assert len(X_gen) == 50
     assert test_plugin.schema_includes(X_gen)
-    assert constraints.filter(X_gen.dataframe()).sum() == len(X_gen)
+    assert constraints.filter(X_gen).sum() == len(X_gen)
     assert list(X_gen.columns) == list(X.columns)
 
 
@@ -131,8 +102,7 @@ def test_sample_hyperparams() -> None:
 
 
 @pytest.mark.slow
-@pytest.mark.parametrize("compress_dataset", [True, False])
-def test_eval_performance(compress_dataset: bool) -> None:
+def test_eval_performance() -> None:
     results = []
 
     Xraw, y = load_iris(return_X_y=True, as_frame=True)
@@ -140,7 +110,9 @@ def test_eval_performance(compress_dataset: bool) -> None:
     X = GenericDataLoader(Xraw)
 
     for retry in range(2):
-        test_plugin = plugin(n_iter=5000, compress_dataset=compress_dataset)
+        test_plugin = plugin(
+            n_iter=200, generator_n_layers_hidden=1, n_teachers=2, lamda=2e-4
+        )
         evaluator = PerformanceEvaluatorXGB()
 
         test_plugin.fit(X)
@@ -148,5 +120,5 @@ def test_eval_performance(compress_dataset: bool) -> None:
 
         results.append(evaluator.evaluate(X, X_syn)["syn_id"])
 
-    print(plugin.name(), results)
-    assert np.mean(results) > 0.8
+    print(plugin.name(), np.mean(results))
+    assert np.mean(results) > 0.7
