+++ tests/plugins/privacy/test_privbayes.py
@@ -1,12 +1,12 @@
 # third party
 import pytest
-from generic_helpers import generate_fixtures
+from fhelpers import generate_fixtures
 from sklearn.datasets import load_iris
 
 # synthcity absolute
 from synthcity.plugins import Plugin
 from synthcity.plugins.core.dataloader import GenericDataLoader
-from synthcity.plugins.generic.plugin_privbayes import plugin
+from synthcity.plugins.privacy.plugin_privbayes import plugin
 
 plugin_name = "privbayes"
 
@@ -23,7 +23,7 @@ def test_plugin_name(test_plugin: Plugin) -> None:
 
 @pytest.mark.parametrize("test_plugin", generate_fixtures(plugin_name, plugin))
 def test_plugin_type(test_plugin: Plugin) -> None:
-    assert test_plugin.type() == "generic"
+    assert test_plugin.type() == "privacy"
 
 
 @pytest.mark.parametrize("test_plugin", generate_fixtures(plugin_name, plugin))
